Braille/backend/requirement.txt
flask==3.0.3
flask-cors==5.0.0
transformers==4.37.2
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.4
